AgenticApp/requirements.txt
openai-agents==0.0.9
pydantic>=2.0.0
python-dotenv>=0.19.0
flask>=2.0.0
